TTG_Backend/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.0.0
numpy>=1.19.0
python-dotenv>=0.19.0
fastapi>=0.95.0
python-multipart>=0.0.6
Pillow>=9.0.0
accelerate>=0.20.0
bitsandbytes>=0.40.0
einops>=0.6.1
peft>=0.4.0
scipy>=1.10.1
timm>=0.6.13
torch-geometric>=2.0.0
networkx>=2.5
matplotlib>=3.3.0
graphviz>=0.16
pytest>=6.0.0
pytest-cov>=2.0.0
protobuf>=3.20.0
sentencepiece --no-build-isolation 